separate customers commands, place db context and decorators in separate folders
